[Compiler] move expressions out of convert file. (#882)
* move expressions out of convert file.

* Apply suggestions from code review

---------
