+++ src/Neo.Compiler.CSharp/MethodConvert/MethodConvert.cs
@@ -935,849 +935,8 @@ namespace Neo.Compiler
             }
         }
 
-        private void ConvertAnonymousObjectCreationExpression(SemanticModel model, AnonymousObjectCreationExpressionSyntax expression)
-        {
-            AddInstruction(OpCode.NEWARRAY0);
-            foreach (AnonymousObjectMemberDeclaratorSyntax initializer in expression.Initializers)
-            {
-                AddInstruction(OpCode.DUP);
-                ConvertExpression(model, initializer.Expression);
-                AddInstruction(OpCode.APPEND);
-            }
-        }
 
-        private void ConvertArrayCreationExpression(SemanticModel model, ArrayCreationExpressionSyntax expression)
-        {
-            ArrayRankSpecifierSyntax specifier = expression.Type.RankSpecifiers[0];
-            if (specifier.Rank != 1)
-                throw new CompilationException(specifier, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {specifier}");
-            IArrayTypeSymbol type = (IArrayTypeSymbol)model.GetTypeInfo(expression.Type).Type!;
-            if (expression.Initializer is null)
-            {
-                ConvertExpression(model, specifier.Sizes[0]);
-                if (type.ElementType.SpecialType == SpecialType.System_Byte)
-                    AddInstruction(OpCode.NEWBUFFER);
-                else
-                    AddInstruction(new Instruction { OpCode = OpCode.NEWARRAY_T, Operand = new[] { (byte)type.ElementType.GetStackItemType() } });
-            }
-            else
-            {
-                ConvertInitializerExpression(model, type, expression.Initializer);
-            }
-        }
 
-        private void ConvertAssignmentExpression(SemanticModel model, AssignmentExpressionSyntax expression)
-        {
-            switch (expression.OperatorToken.ValueText)
-            {
-                case "=":
-                    ConvertSimpleAssignmentExpression(model, expression);
-                    break;
-                case "??=":
-                    ConvertCoalesceAssignmentExpression(model, expression);
-                    break;
-                default:
-                    ConvertComplexAssignmentExpression(model, expression);
-                    break;
-            }
-        }
-
-        private void ConvertSimpleAssignmentExpression(SemanticModel model, AssignmentExpressionSyntax expression)
-        {
-            ConvertExpression(model, expression.Right);
-            AddInstruction(OpCode.DUP);
-            switch (expression.Left)
-            {
-                case DeclarationExpressionSyntax left:
-                    ConvertDeclarationAssignment(model, left);
-                    break;
-                case ElementAccessExpressionSyntax left:
-                    ConvertElementAccessAssignment(model, left);
-                    break;
-                case IdentifierNameSyntax left:
-                    ConvertIdentifierNameAssignment(model, left);
-                    break;
-                case MemberAccessExpressionSyntax left:
-                    ConvertMemberAccessAssignment(model, left);
-                    break;
-                case TupleExpressionSyntax left:
-                    ConvertTupleAssignment(model, left);
-                    break;
-                default:
-                    throw new CompilationException(expression.Left, DiagnosticId.SyntaxNotSupported, $"Unsupported assignment: {expression.Left}");
-            }
-        }
-
-        private void ConvertDeclarationAssignment(SemanticModel model, DeclarationExpressionSyntax left)
-        {
-            ITypeSymbol type = model.GetTypeInfo(left).Type!;
-            if (!type.IsValueType)
-                throw new CompilationException(left, DiagnosticId.SyntaxNotSupported, $"Unsupported assignment type: {type}");
-            AddInstruction(OpCode.UNPACK);
-            AddInstruction(OpCode.DROP);
-            foreach (VariableDesignationSyntax variable in ((ParenthesizedVariableDesignationSyntax)left.Designation).Variables)
-            {
-                switch (variable)
-                {
-                    case SingleVariableDesignationSyntax singleVariableDesignation:
-                        ILocalSymbol local = (ILocalSymbol)model.GetDeclaredSymbol(singleVariableDesignation)!;
-                        byte index = AddLocalVariable(local);
-                        AccessSlot(OpCode.STLOC, index);
-                        break;
-                    case DiscardDesignationSyntax:
-                        AddInstruction(OpCode.DROP);
-                        break;
-                    default:
-                        throw new CompilationException(variable, DiagnosticId.SyntaxNotSupported, $"Unsupported designation: {variable}");
-                }
-            }
-        }
-
-        private void ConvertElementAccessAssignment(SemanticModel model, ElementAccessExpressionSyntax left)
-        {
-            if (left.ArgumentList.Arguments.Count != 1)
-                throw new CompilationException(left.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {left.ArgumentList.Arguments}");
-            if (model.GetSymbolInfo(left).Symbol is IPropertySymbol property)
-            {
-                ConvertExpression(model, left.ArgumentList.Arguments[0].Expression);
-                ConvertExpression(model, left.Expression);
-                Call(model, property.SetMethod!, CallingConvention.Cdecl);
-            }
-            else
-            {
-                ConvertExpression(model, left.Expression);
-                ConvertExpression(model, left.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.ROT);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertIdentifierNameAssignment(SemanticModel model, IdentifierNameSyntax left)
-        {
-            ISymbol symbol = model.GetSymbolInfo(left).Symbol!;
-            switch (symbol)
-            {
-                case IDiscardSymbol:
-                    AddInstruction(OpCode.DROP);
-                    break;
-                case IFieldSymbol field:
-                    if (field.IsStatic)
-                    {
-                        byte index = context.AddStaticField(field);
-                        AccessSlot(OpCode.STSFLD, index);
-                    }
-                    else
-                    {
-                        int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                        AddInstruction(OpCode.LDARG0);
-                        Push(index);
-                        AddInstruction(OpCode.ROT);
-                        AddInstruction(OpCode.SETITEM);
-                    }
-                    break;
-                case ILocalSymbol local:
-                    AccessSlot(OpCode.STLOC, _localVariables[local]);
-                    break;
-                case IParameterSymbol parameter:
-                    AccessSlot(OpCode.STARG, _parameters[parameter]);
-                    break;
-                case IPropertySymbol property:
-                    if (!property.IsStatic) AddInstruction(OpCode.LDARG0);
-                    Call(model, property.SetMethod!, CallingConvention.Cdecl);
-                    break;
-                default:
-                    throw new CompilationException(left, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertMemberAccessAssignment(SemanticModel model, MemberAccessExpressionSyntax left)
-        {
-            ISymbol symbol = model.GetSymbolInfo(left.Name).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    if (field.IsStatic)
-                    {
-                        byte index = context.AddStaticField(field);
-                        AccessSlot(OpCode.STSFLD, index);
-                    }
-                    else
-                    {
-                        int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                        ConvertExpression(model, left.Expression);
-                        Push(index);
-                        AddInstruction(OpCode.ROT);
-                        AddInstruction(OpCode.SETITEM);
-                    }
-                    break;
-                case IPropertySymbol property:
-                    if (!property.IsStatic) ConvertExpression(model, left.Expression);
-                    Call(model, property.SetMethod!, CallingConvention.Cdecl);
-                    break;
-                default:
-                    throw new CompilationException(left, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertTupleAssignment(SemanticModel model, TupleExpressionSyntax left)
-        {
-            AddInstruction(OpCode.UNPACK);
-            AddInstruction(OpCode.DROP);
-            foreach (ArgumentSyntax argument in left.Arguments)
-            {
-                switch (argument.Expression)
-                {
-                    case DeclarationExpressionSyntax declaration:
-                        switch (declaration.Designation)
-                        {
-                            case SingleVariableDesignationSyntax singleVariableDesignation:
-                                ILocalSymbol local = (ILocalSymbol)model.GetDeclaredSymbol(singleVariableDesignation)!;
-                                byte index = AddLocalVariable(local);
-                                AccessSlot(OpCode.STLOC, index);
-                                break;
-                            case DiscardDesignationSyntax:
-                                AddInstruction(OpCode.DROP);
-                                break;
-                            default:
-                                throw new CompilationException(argument, DiagnosticId.SyntaxNotSupported, $"Unsupported designation: {argument}");
-                        }
-                        break;
-                    case IdentifierNameSyntax identifier:
-                        ConvertIdentifierNameAssignment(model, identifier);
-                        break;
-                    case MemberAccessExpressionSyntax memberAccess:
-                        ConvertMemberAccessAssignment(model, memberAccess);
-                        break;
-                    default:
-                        throw new CompilationException(argument, DiagnosticId.SyntaxNotSupported, $"Unsupported assignment: {argument}");
-                }
-            }
-        }
-
-        private void ConvertCoalesceAssignmentExpression(SemanticModel model, AssignmentExpressionSyntax expression)
-        {
-            switch (expression.Left)
-            {
-                case ElementAccessExpressionSyntax left:
-                    ConvertElementAccessCoalesceAssignment(model, left, expression.Right);
-                    break;
-                case IdentifierNameSyntax left:
-                    ConvertIdentifierNameCoalesceAssignment(model, left, expression.Right);
-                    break;
-                case MemberAccessExpressionSyntax left:
-                    ConvertMemberAccessCoalesceAssignment(model, left, expression.Right);
-                    break;
-                default:
-                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported coalesce assignment: {expression}");
-            }
-        }
-
-        private void ConvertElementAccessCoalesceAssignment(SemanticModel model, ElementAccessExpressionSyntax left, ExpressionSyntax right)
-        {
-            if (left.ArgumentList.Arguments.Count != 1)
-                throw new CompilationException(left.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {left.ArgumentList.Arguments}");
-            JumpTarget assignmentTarget = new();
-            JumpTarget endTarget = new();
-            if (model.GetSymbolInfo(left).Symbol is IPropertySymbol property)
-            {
-                ConvertExpression(model, left.Expression);
-                ConvertExpression(model, left.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                Call(model, property.GetMethod!, CallingConvention.StdCall);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIF_L, assignmentTarget);
-                AddInstruction(OpCode.NIP);
-                AddInstruction(OpCode.NIP);
-                Jump(OpCode.JMP_L, endTarget);
-                assignmentTarget.Instruction = AddInstruction(OpCode.DROP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                Call(model, property.SetMethod!, CallingConvention.Cdecl);
-            }
-            else
-            {
-                ConvertExpression(model, left.Expression);
-                ConvertExpression(model, left.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.PICKITEM);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIF_L, assignmentTarget);
-                AddInstruction(OpCode.PICKITEM);
-                Jump(OpCode.JMP_L, endTarget);
-                assignmentTarget.Instruction = AddInstruction(OpCode.NOP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                AddInstruction(OpCode.REVERSE3);
-                AddInstruction(OpCode.SETITEM);
-            }
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertIdentifierNameCoalesceAssignment(SemanticModel model, IdentifierNameSyntax left, ExpressionSyntax right)
-        {
-            ISymbol symbol = model.GetSymbolInfo(left).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldIdentifierNameCoalesceAssignment(model, field, right);
-                    break;
-                case ILocalSymbol local:
-                    ConvertLocalIdentifierNameCoalesceAssignment(model, local, right);
-                    break;
-                case IParameterSymbol parameter:
-                    ConvertParameterIdentifierNameCoalesceAssignment(model, parameter, right);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyIdentifierNameCoalesceAssignment(model, property, right);
-                    break;
-                default:
-                    throw new CompilationException(left, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldIdentifierNameCoalesceAssignment(SemanticModel model, IFieldSymbol left, ExpressionSyntax right)
-        {
-            JumpTarget assignmentTarget = new();
-            JumpTarget endTarget = new();
-            if (left.IsStatic)
-            {
-                byte index = context.AddStaticField(left);
-                AccessSlot(OpCode.LDSFLD, index);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIF_L, assignmentTarget);
-                AccessSlot(OpCode.LDSFLD, index);
-                Jump(OpCode.JMP_L, endTarget);
-                assignmentTarget.Instruction = AddInstruction(OpCode.NOP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(left.ContainingType.GetFields(), left);
-                AddInstruction(OpCode.LDARG0);
-                Push(index);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.PICKITEM);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIF_L, assignmentTarget);
-                AddInstruction(OpCode.PICKITEM);
-                Jump(OpCode.JMP_L, endTarget);
-                assignmentTarget.Instruction = AddInstruction(OpCode.NOP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                AddInstruction(OpCode.REVERSE3);
-                AddInstruction(OpCode.SETITEM);
-            }
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertLocalIdentifierNameCoalesceAssignment(SemanticModel model, ILocalSymbol left, ExpressionSyntax right)
-        {
-            JumpTarget assignmentTarget = new();
-            JumpTarget endTarget = new();
-            byte index = _localVariables[left];
-            AccessSlot(OpCode.LDLOC, index);
-            AddInstruction(OpCode.ISNULL);
-            Jump(OpCode.JMPIF_L, assignmentTarget);
-            AccessSlot(OpCode.LDLOC, index);
-            Jump(OpCode.JMP_L, endTarget);
-            assignmentTarget.Instruction = AddInstruction(OpCode.NOP);
-            ConvertExpression(model, right);
-            AddInstruction(OpCode.DUP);
-            AccessSlot(OpCode.STLOC, index);
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertParameterIdentifierNameCoalesceAssignment(SemanticModel model, IParameterSymbol left, ExpressionSyntax right)
-        {
-            JumpTarget assignmentTarget = new();
-            JumpTarget endTarget = new();
-            byte index = _parameters[left];
-            AccessSlot(OpCode.LDARG, index);
-            AddInstruction(OpCode.ISNULL);
-            Jump(OpCode.JMPIF_L, assignmentTarget);
-            AccessSlot(OpCode.LDARG, index);
-            Jump(OpCode.JMP_L, endTarget);
-            assignmentTarget.Instruction = AddInstruction(OpCode.NOP);
-            ConvertExpression(model, right);
-            AddInstruction(OpCode.DUP);
-            AccessSlot(OpCode.STARG, index);
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertPropertyIdentifierNameCoalesceAssignment(SemanticModel model, IPropertySymbol left, ExpressionSyntax right)
-        {
-            JumpTarget endTarget = new();
-            if (left.IsStatic)
-            {
-                Call(model, left.GetMethod!);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIFNOT_L, endTarget);
-                AddInstruction(OpCode.DROP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                Call(model, left.SetMethod!);
-            }
-            else
-            {
-                AddInstruction(OpCode.LDARG0);
-                Call(model, left.GetMethod!);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIFNOT_L, endTarget);
-                AddInstruction(OpCode.DROP);
-                AddInstruction(OpCode.LDARG0);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.TUCK);
-                Call(model, left.SetMethod!, CallingConvention.StdCall);
-            }
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertMemberAccessCoalesceAssignment(SemanticModel model, MemberAccessExpressionSyntax left, ExpressionSyntax right)
-        {
-            ISymbol symbol = model.GetSymbolInfo(left).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldMemberAccessCoalesceAssignment(model, left, right, field);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyMemberAccessCoalesceAssignment(model, left, right, property);
-                    break;
-                default:
-                    throw new CompilationException(left, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldMemberAccessCoalesceAssignment(SemanticModel model, MemberAccessExpressionSyntax left, ExpressionSyntax right, IFieldSymbol field)
-        {
-            JumpTarget assignmentTarget = new();
-            JumpTarget endTarget = new();
-            if (field.IsStatic)
-            {
-                byte index = context.AddStaticField(field);
-                AccessSlot(OpCode.LDSFLD, index);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIF_L, assignmentTarget);
-                AccessSlot(OpCode.LDSFLD, index);
-                Jump(OpCode.JMP_L, endTarget);
-                assignmentTarget.Instruction = AddInstruction(OpCode.NOP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                ConvertExpression(model, left.Expression);
-                Push(index);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.PICKITEM);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIF_L, assignmentTarget);
-                AddInstruction(OpCode.PICKITEM);
-                Jump(OpCode.JMP_L, endTarget);
-                assignmentTarget.Instruction = AddInstruction(OpCode.NOP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                AddInstruction(OpCode.REVERSE3);
-                AddInstruction(OpCode.SETITEM);
-            }
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertPropertyMemberAccessCoalesceAssignment(SemanticModel model, MemberAccessExpressionSyntax left, ExpressionSyntax right, IPropertySymbol property)
-        {
-            JumpTarget endTarget = new();
-            if (property.IsStatic)
-            {
-                Call(model, property.GetMethod!);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIFNOT_L, endTarget);
-                AddInstruction(OpCode.DROP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.DUP);
-                Call(model, property.SetMethod!);
-            }
-            else
-            {
-                JumpTarget assignmentTarget = new();
-                ConvertExpression(model, left.Expression);
-                AddInstruction(OpCode.DUP);
-                Call(model, property.GetMethod!);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIF_L, assignmentTarget);
-                AddInstruction(OpCode.NIP);
-                Jump(OpCode.JMP_L, endTarget);
-                assignmentTarget.Instruction = AddInstruction(OpCode.DROP);
-                ConvertExpression(model, right);
-                AddInstruction(OpCode.TUCK);
-                Call(model, property.SetMethod!, CallingConvention.StdCall);
-            }
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertComplexAssignmentExpression(SemanticModel model, AssignmentExpressionSyntax expression)
-        {
-            ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-            switch (expression.Left)
-            {
-                case ElementAccessExpressionSyntax left:
-                    ConvertElementAccessComplexAssignment(model, type, expression.OperatorToken, left, expression.Right);
-                    break;
-                case IdentifierNameSyntax left:
-                    ConvertIdentifierNameComplexAssignment(model, type, expression.OperatorToken, left, expression.Right);
-                    break;
-                case MemberAccessExpressionSyntax left:
-                    ConvertMemberAccessComplexAssignment(model, type, expression.OperatorToken, left, expression.Right);
-                    break;
-                default:
-                    throw new CompilationException(expression.Left, DiagnosticId.SyntaxNotSupported, $"Unsupported assignment expression: {expression}");
-            }
-        }
-
-        private void ConvertElementAccessComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, ElementAccessExpressionSyntax left, ExpressionSyntax right)
-        {
-            if (left.ArgumentList.Arguments.Count != 1)
-                throw new CompilationException(left.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {left.ArgumentList.Arguments}");
-            if (model.GetSymbolInfo(left).Symbol is IPropertySymbol property)
-            {
-                ConvertExpression(model, left.Expression);
-                ConvertExpression(model, left.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                Call(model, property.GetMethod!, CallingConvention.StdCall);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                Call(model, property.SetMethod!, CallingConvention.Cdecl);
-            }
-            else
-            {
-                ConvertExpression(model, left.Expression);
-                ConvertExpression(model, left.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.PICKITEM);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                AddInstruction(OpCode.REVERSE3);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertIdentifierNameComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, IdentifierNameSyntax left, ExpressionSyntax right)
-        {
-            ISymbol symbol = model.GetSymbolInfo(left).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldIdentifierNameComplexAssignment(model, type, operatorToken, field, right);
-                    break;
-                case ILocalSymbol local:
-                    ConvertLocalIdentifierNameComplexAssignment(model, type, operatorToken, local, right);
-                    break;
-                case IParameterSymbol parameter:
-                    ConvertParameterIdentifierNameComplexAssignment(model, type, operatorToken, parameter, right);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyIdentifierNameComplexAssignment(model, type, operatorToken, property, right);
-                    break;
-                default:
-                    throw new CompilationException(left, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldIdentifierNameComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, IFieldSymbol left, ExpressionSyntax right)
-        {
-            if (left.IsStatic)
-            {
-                byte index = context.AddStaticField(left);
-                AccessSlot(OpCode.LDSFLD, index);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(left.ContainingType.GetFields(), left);
-                AddInstruction(OpCode.LDARG0);
-                AddInstruction(OpCode.DUP);
-                Push(index);
-                AddInstruction(OpCode.PICKITEM);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.TUCK);
-                Push(index);
-                AddInstruction(OpCode.SWAP);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertLocalIdentifierNameComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, ILocalSymbol left, ExpressionSyntax right)
-        {
-            byte index = _localVariables[left];
-            AccessSlot(OpCode.LDLOC, index);
-            ConvertExpression(model, right);
-            EmitComplexAssignmentOperator(type, operatorToken);
-            AddInstruction(OpCode.DUP);
-            AccessSlot(OpCode.STLOC, index);
-        }
-
-        private void ConvertParameterIdentifierNameComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, IParameterSymbol left, ExpressionSyntax right)
-        {
-            byte index = _parameters[left];
-            AccessSlot(OpCode.LDARG, index);
-            ConvertExpression(model, right);
-            EmitComplexAssignmentOperator(type, operatorToken);
-            AddInstruction(OpCode.DUP);
-            AccessSlot(OpCode.STARG, index);
-        }
-
-        private void ConvertPropertyIdentifierNameComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, IPropertySymbol left, ExpressionSyntax right)
-        {
-            if (left.IsStatic)
-            {
-                Call(model, left.GetMethod!);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.DUP);
-                Call(model, left.SetMethod!);
-            }
-            else
-            {
-                AddInstruction(OpCode.LDARG0);
-                AddInstruction(OpCode.DUP);
-                Call(model, left.GetMethod!);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.TUCK);
-                Call(model, left.SetMethod!, CallingConvention.StdCall);
-            }
-        }
-
-        private void ConvertMemberAccessComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, MemberAccessExpressionSyntax left, ExpressionSyntax right)
-        {
-            ISymbol symbol = model.GetSymbolInfo(left).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldMemberAccessComplexAssignment(model, type, operatorToken, left, right, field);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyMemberAccessComplexAssignment(model, type, operatorToken, left, right, property);
-                    break;
-                default:
-                    throw new CompilationException(left, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldMemberAccessComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, MemberAccessExpressionSyntax left, ExpressionSyntax right, IFieldSymbol field)
-        {
-            if (field.IsStatic)
-            {
-                byte index = context.AddStaticField(field);
-                AccessSlot(OpCode.LDSFLD, index);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                ConvertExpression(model, left.Expression);
-                AddInstruction(OpCode.DUP);
-                Push(index);
-                AddInstruction(OpCode.PICKITEM);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.TUCK);
-                Push(index);
-                AddInstruction(OpCode.SWAP);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertPropertyMemberAccessComplexAssignment(SemanticModel model, ITypeSymbol type, SyntaxToken operatorToken, MemberAccessExpressionSyntax left, ExpressionSyntax right, IPropertySymbol property)
-        {
-            if (property.IsStatic)
-            {
-                Call(model, property.GetMethod!);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.DUP);
-                Call(model, property.SetMethod!);
-            }
-            else
-            {
-                ConvertExpression(model, left.Expression);
-                AddInstruction(OpCode.DUP);
-                Call(model, property.GetMethod!);
-                ConvertExpression(model, right);
-                EmitComplexAssignmentOperator(type, operatorToken);
-                AddInstruction(OpCode.TUCK);
-                Call(model, property.SetMethod!, CallingConvention.StdCall);
-            }
-        }
-
-        private void EmitComplexAssignmentOperator(ITypeSymbol type, SyntaxToken operatorToken)
-        {
-            var itemType = type.GetStackItemType();
-            bool isBoolean = itemType == VM.Types.StackItemType.Boolean;
-            bool isString = itemType == VM.Types.StackItemType.ByteString;
-
-            var (opcode, checkResult) = operatorToken.ValueText switch
-            {
-                "+=" => isString ? (OpCode.CAT, false) : (OpCode.ADD, true),
-                "-=" => (OpCode.SUB, true),
-                "*=" => (OpCode.MUL, true),
-                "/=" => (OpCode.DIV, true),
-                "%=" => (OpCode.MOD, true),
-                "&=" => isBoolean ? (OpCode.BOOLAND, false) : (OpCode.AND, true),
-                "^=" when !isBoolean => (OpCode.XOR, true),
-                "|=" => isBoolean ? (OpCode.BOOLOR, false) : (OpCode.OR, true),
-                "<<=" => (OpCode.SHL, true),
-                ">>=" => (OpCode.SHR, true),
-                _ => throw new CompilationException(operatorToken, DiagnosticId.SyntaxNotSupported, $"Unsupported operator: {operatorToken}")
-            };
-            AddInstruction(opcode);
-            if (isString) ChangeType(VM.Types.StackItemType.ByteString);
-            if (checkResult) EnsureIntegerInRange(type);
-        }
-
-        private void ConvertObjectCreationExpression(SemanticModel model, BaseObjectCreationExpressionSyntax expression)
-        {
-            ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-            if (type.TypeKind == TypeKind.Delegate)
-            {
-                ConvertDelegateCreationExpression(model, expression);
-                return;
-            }
-            IMethodSymbol constructor = (IMethodSymbol)model.GetSymbolInfo(expression).Symbol!;
-            IReadOnlyList<ArgumentSyntax> arguments = expression.ArgumentList?.Arguments ?? (IReadOnlyList<ArgumentSyntax>)Array.Empty<ArgumentSyntax>();
-            if (TryProcessSystemConstructors(model, constructor, arguments))
-                return;
-            bool needCreateObject = !type.DeclaringSyntaxReferences.IsEmpty && !constructor.IsExtern;
-            if (needCreateObject)
-            {
-                CreateObject(model, type, null);
-            }
-            Call(model, constructor, needCreateObject, arguments);
-            if (expression.Initializer is not null)
-            {
-                ConvertObjectCreationExpressionInitializer(model, expression.Initializer);
-            }
-        }
-
-        private void ConvertObjectCreationExpressionInitializer(SemanticModel model, InitializerExpressionSyntax initializer)
-        {
-            foreach (ExpressionSyntax e in initializer.Expressions)
-            {
-                if (e is not AssignmentExpressionSyntax ae)
-                    throw new CompilationException(initializer, DiagnosticId.SyntaxNotSupported, $"Unsupported initializer: {initializer}");
-                ISymbol symbol = model.GetSymbolInfo(ae.Left).Symbol!;
-                switch (symbol)
-                {
-                    case IFieldSymbol field:
-                        AddInstruction(OpCode.DUP);
-                        int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                        Push(index);
-                        ConvertExpression(model, ae.Right);
-                        AddInstruction(OpCode.SETITEM);
-                        break;
-                    case IPropertySymbol property:
-                        ConvertExpression(model, ae.Right);
-                        AddInstruction(OpCode.OVER);
-                        Call(model, property.SetMethod!, CallingConvention.Cdecl);
-                        break;
-                    default:
-                        throw new CompilationException(ae.Left, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-                }
-            }
-        }
-
-        private void ConvertDelegateCreationExpression(SemanticModel model, BaseObjectCreationExpressionSyntax expression)
-        {
-            if (expression.ArgumentList!.Arguments.Count != 1)
-                throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported delegate: {expression}");
-            IMethodSymbol symbol = (IMethodSymbol)model.GetSymbolInfo(expression.ArgumentList.Arguments[0].Expression).Symbol!;
-            if (!symbol.IsStatic)
-                throw new CompilationException(expression, DiagnosticId.NonStaticDelegate, $"Unsupported delegate: {symbol}");
-            MethodConvert convert = context.ConvertMethod(model, symbol);
-            Jump(OpCode.PUSHA, convert._startTarget);
-        }
-
-        private void ConvertBinaryExpression(SemanticModel model, BinaryExpressionSyntax expression)
-        {
-            switch (expression.OperatorToken.ValueText)
-            {
-                case "||":
-                    ConvertLogicalOrExpression(model, expression.Left, expression.Right);
-                    return;
-                case "&&":
-                    ConvertLogicalAndExpression(model, expression.Left, expression.Right);
-                    return;
-                case "is":
-                    ConvertIsExpression(model, expression.Left, expression.Right);
-                    return;
-                case "as":
-                    ConvertAsExpression(model, expression.Left, expression.Right);
-                    return;
-                case "??":
-                    ConvertCoalesceExpression(model, expression.Left, expression.Right);
-                    return;
-            }
-            IMethodSymbol? symbol = (IMethodSymbol?)model.GetSymbolInfo(expression).Symbol;
-            if (symbol is not null && TryProcessSystemOperators(model, symbol, expression.Left, expression.Right))
-                return;
-            ConvertExpression(model, expression.Left);
-            ConvertExpression(model, expression.Right);
-            (OpCode opcode, bool checkResult) = expression.OperatorToken.ValueText switch
-            {
-                "+" => (OpCode.ADD, true),
-                "-" => (OpCode.SUB, true),
-                "*" => (OpCode.MUL, true),
-                "/" => (OpCode.DIV, false),
-                "%" => (OpCode.MOD, false),
-                "<<" => (OpCode.SHL, true),
-                ">>" => (OpCode.SHR, false),
-                "|" => (OpCode.OR, false),
-                "&" => (OpCode.AND, false),
-                "^" => (OpCode.XOR, false),
-                "==" => (OpCode.EQUAL, false),
-                "!=" => (OpCode.NOTEQUAL, false),
-                "<" => (OpCode.LT, false),
-                "<=" => (OpCode.LE, false),
-                ">" => (OpCode.GT, false),
-                ">=" => (OpCode.GE, false),
-                _ => throw new CompilationException(expression.OperatorToken, DiagnosticId.SyntaxNotSupported, $"Unsupported operator: {expression.OperatorToken}")
-            };
-            AddInstruction(opcode);
-            if (checkResult)
-            {
-                ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-                EnsureIntegerInRange(type);
-            }
-        }
 
         private void EnsureIntegerInRange(ITypeSymbol type)
         {
@@ -1825,408 +984,6 @@ namespace Neo.Compiler
             endTarget.Instruction = AddInstruction(OpCode.NOP);
         }
 
-        private void ConvertLogicalOrExpression(SemanticModel model, ExpressionSyntax left, ExpressionSyntax right)
-        {
-            JumpTarget rightTarget = new();
-            JumpTarget endTarget = new();
-            ConvertExpression(model, left);
-            Jump(OpCode.JMPIFNOT_L, rightTarget);
-            Push(true);
-            Jump(OpCode.JMP_L, endTarget);
-            rightTarget.Instruction = AddInstruction(OpCode.NOP);
-            ConvertExpression(model, right);
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertLogicalAndExpression(SemanticModel model, ExpressionSyntax left, ExpressionSyntax right)
-        {
-            JumpTarget rightTarget = new();
-            JumpTarget endTarget = new();
-            ConvertExpression(model, left);
-            Jump(OpCode.JMPIF_L, rightTarget);
-            Push(false);
-            Jump(OpCode.JMP_L, endTarget);
-            rightTarget.Instruction = AddInstruction(OpCode.NOP);
-            ConvertExpression(model, right);
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertIsExpression(SemanticModel model, ExpressionSyntax left, ExpressionSyntax right)
-        {
-            ITypeSymbol type = model.GetTypeInfo(right).Type!;
-            ConvertExpression(model, left);
-            IsType(type.GetPatternType());
-        }
-
-        private void ConvertAsExpression(SemanticModel model, ExpressionSyntax left, ExpressionSyntax right)
-        {
-            JumpTarget endTarget = new();
-            ITypeSymbol type = model.GetTypeInfo(right).Type!;
-            ConvertExpression(model, left);
-            AddInstruction(OpCode.DUP);
-            IsType(type.GetPatternType());
-            Jump(OpCode.JMPIF_L, endTarget);
-            AddInstruction(OpCode.DROP);
-            Push((object?)null);
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertCoalesceExpression(SemanticModel model, ExpressionSyntax left, ExpressionSyntax right)
-        {
-            JumpTarget endTarget = new();
-            ConvertExpression(model, left);
-            AddInstruction(OpCode.DUP);
-            AddInstruction(OpCode.ISNULL);
-            Jump(OpCode.JMPIFNOT_L, endTarget);
-            AddInstruction(OpCode.DROP);
-            ConvertExpression(model, right);
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertCastExpression(SemanticModel model, CastExpressionSyntax expression)
-        {
-            ITypeSymbol sType = model.GetTypeInfo(expression.Expression).Type!;
-            ITypeSymbol tType = model.GetTypeInfo(expression.Type).Type!;
-            IMethodSymbol method = (IMethodSymbol)model.GetSymbolInfo(expression).Symbol!;
-            if (method is not null)
-            {
-                Call(model, method, null, expression.Expression);
-                return;
-            }
-            ConvertExpression(model, expression.Expression);
-            switch ((sType.Name, tType.Name))
-            {
-                case ("ByteString", "ECPoint"):
-                    {
-                        JumpTarget endTarget = new();
-                        AddInstruction(OpCode.DUP);
-                        AddInstruction(OpCode.ISNULL);
-                        Jump(OpCode.JMPIF_L, endTarget);
-                        AddInstruction(OpCode.DUP);
-                        AddInstruction(OpCode.SIZE);
-                        Push(33);
-                        Jump(OpCode.JMPEQ_L, endTarget);
-                        AddInstruction(OpCode.THROW);
-                        endTarget.Instruction = AddInstruction(OpCode.NOP);
-                    }
-                    break;
-                case ("ByteString", "UInt160"):
-                    {
-                        JumpTarget endTarget = new();
-                        AddInstruction(OpCode.DUP);
-                        AddInstruction(OpCode.ISNULL);
-                        Jump(OpCode.JMPIF_L, endTarget);
-                        AddInstruction(OpCode.DUP);
-                        AddInstruction(OpCode.SIZE);
-                        Push(20);
-                        Jump(OpCode.JMPEQ_L, endTarget);
-                        AddInstruction(OpCode.THROW);
-                        endTarget.Instruction = AddInstruction(OpCode.NOP);
-                    }
-                    break;
-                case ("ByteString", "UInt256"):
-                    {
-                        JumpTarget endTarget = new();
-                        AddInstruction(OpCode.DUP);
-                        AddInstruction(OpCode.ISNULL);
-                        Jump(OpCode.JMPIF_L, endTarget);
-                        AddInstruction(OpCode.DUP);
-                        AddInstruction(OpCode.SIZE);
-                        Push(32);
-                        Jump(OpCode.JMPEQ_L, endTarget);
-                        AddInstruction(OpCode.THROW);
-                        endTarget.Instruction = AddInstruction(OpCode.NOP);
-                    }
-                    break;
-                case ("SByte", "Byte"):
-                case ("SByte", "UInt16"):
-                case ("SByte", "UInt32"):
-                case ("SByte", "UInt64"):
-                case ("Int16", "SByte"):
-                case ("Int16", "Byte"):
-                case ("Int16", "UInt16"):
-                case ("Int16", "UInt32"):
-                case ("Int16", "UInt64"):
-                case ("Int32", "SByte"):
-                case ("Int32", "Int16"):
-                case ("Int32", "Byte"):
-                case ("Int32", "UInt16"):
-                case ("Int32", "UInt32"):
-                case ("Int32", "UInt64"):
-                case ("Int64", "SByte"):
-                case ("Int64", "Int16"):
-                case ("Int64", "Int32"):
-                case ("Int64", "Byte"):
-                case ("Int64", "UInt16"):
-                case ("Int64", "UInt32"):
-                case ("Int64", "UInt64"):
-                case ("Byte", "SByte"):
-                case ("UInt16", "SByte"):
-                case ("UInt16", "Int16"):
-                case ("UInt16", "Byte"):
-                case ("UInt32", "SByte"):
-                case ("UInt32", "Int16"):
-                case ("UInt32", "Int32"):
-                case ("UInt32", "Byte"):
-                case ("UInt32", "UInt16"):
-                case ("UInt64", "SByte"):
-                case ("UInt64", "Int16"):
-                case ("UInt64", "Int32"):
-                case ("UInt64", "Int64"):
-                case ("UInt64", "Byte"):
-                case ("UInt64", "UInt16"):
-                case ("UInt64", "UInt32"):
-                    {
-                        EnsureIntegerInRange(tType);
-                    }
-                    break;
-            }
-        }
-
-        private void ConvertCheckedExpression(SemanticModel model, CheckedExpressionSyntax expression)
-        {
-            _checkedStack.Push(expression.Keyword.IsKind(SyntaxKind.CheckedKeyword));
-            ConvertExpression(model, expression.Expression);
-            _checkedStack.Pop();
-        }
-
-        private void ConvertConditionalAccessExpression(SemanticModel model, ConditionalAccessExpressionSyntax expression)
-        {
-            ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-            JumpTarget nullTarget = new();
-            ConvertExpression(model, expression.Expression);
-            AddInstruction(OpCode.DUP);
-            AddInstruction(OpCode.ISNULL);
-            Jump(OpCode.JMPIF_L, nullTarget);
-            ConvertExpression(model, expression.WhenNotNull);
-            if (type.SpecialType == SpecialType.System_Void)
-            {
-                JumpTarget endTarget = new();
-                Jump(OpCode.JMP_L, endTarget);
-                nullTarget.Instruction = AddInstruction(OpCode.DROP);
-                endTarget.Instruction = AddInstruction(OpCode.NOP);
-            }
-            else
-            {
-                nullTarget.Instruction = AddInstruction(OpCode.NOP);
-            }
-        }
-
-        private void ConvertConditionalExpression(SemanticModel model, ConditionalExpressionSyntax expression)
-        {
-            JumpTarget falseTarget = new();
-            JumpTarget endTarget = new();
-            ConvertExpression(model, expression.Condition);
-            Jump(OpCode.JMPIFNOT_L, falseTarget);
-            ConvertExpression(model, expression.WhenTrue);
-            Jump(OpCode.JMP_L, endTarget);
-            falseTarget.Instruction = AddInstruction(OpCode.NOP);
-            ConvertExpression(model, expression.WhenFalse);
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertElementAccessExpression(SemanticModel model, ElementAccessExpressionSyntax expression)
-        {
-            if (expression.ArgumentList.Arguments.Count != 1)
-                throw new CompilationException(expression.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {expression.ArgumentList.Arguments}");
-            if (model.GetSymbolInfo(expression).Symbol is IPropertySymbol property)
-            {
-                Call(model, property.GetMethod!, expression.Expression, expression.ArgumentList.Arguments.ToArray());
-            }
-            else
-            {
-                ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-                ConvertExpression(model, expression.Expression);
-                ConvertIndexOrRange(model, type, expression.ArgumentList.Arguments[0].Expression);
-            }
-        }
-
-        private void ConvertElementBindingExpression(SemanticModel model, ElementBindingExpressionSyntax expression)
-        {
-            if (expression.ArgumentList.Arguments.Count != 1)
-                throw new CompilationException(expression.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {expression.ArgumentList.Arguments}");
-            ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-            ConvertIndexOrRange(model, type, expression.ArgumentList.Arguments[0].Expression);
-        }
-
-        private void ConvertIndexOrRange(SemanticModel model, ITypeSymbol type, ExpressionSyntax indexOrRange)
-        {
-            if (indexOrRange is RangeExpressionSyntax range)
-            {
-                if (range.RightOperand is null)
-                {
-                    AddInstruction(OpCode.DUP);
-                    AddInstruction(OpCode.SIZE);
-                }
-                else
-                {
-                    ConvertExpression(model, range.RightOperand);
-                }
-                AddInstruction(OpCode.SWAP);
-                if (range.LeftOperand is null)
-                {
-                    Push(0);
-                }
-                else
-                {
-                    ConvertExpression(model, range.LeftOperand);
-                }
-                AddInstruction(OpCode.ROT);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.SUB);
-                switch (type.ToString())
-                {
-                    case "byte[]":
-                        AddInstruction(OpCode.SUBSTR);
-                        break;
-                    case "string":
-                        AddInstruction(OpCode.SUBSTR);
-                        ChangeType(VM.Types.StackItemType.ByteString);
-                        break;
-                    default:
-                        throw new CompilationException(indexOrRange, DiagnosticId.ArrayRange, $"The type {type} does not support range access.");
-                }
-            }
-            else
-            {
-                ConvertExpression(model, indexOrRange);
-                AddInstruction(OpCode.PICKITEM);
-            }
-        }
-
-        private void ConvertIdentifierNameExpression(SemanticModel model, IdentifierNameSyntax expression)
-        {
-            ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    if (field.IsConst)
-                    {
-                        Push(field.ConstantValue);
-                    }
-                    else if (field.IsStatic)
-                    {
-                        byte index = context.AddStaticField(field);
-                        AccessSlot(OpCode.LDSFLD, index);
-                    }
-                    else
-                    {
-                        int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                        AddInstruction(OpCode.LDARG0);
-                        Push(index);
-                        AddInstruction(OpCode.PICKITEM);
-                    }
-                    break;
-                case ILocalSymbol local:
-                    if (local.IsConst)
-                        Push(local.ConstantValue);
-                    else
-                        AccessSlot(OpCode.LDLOC, _localVariables[local]);
-                    break;
-                case IMethodSymbol method:
-                    if (!method.IsStatic)
-                        throw new CompilationException(expression, DiagnosticId.NonStaticDelegate, $"Unsupported delegate: {method}");
-                    MethodConvert convert = context.ConvertMethod(model, method);
-                    Jump(OpCode.PUSHA, convert._startTarget);
-                    break;
-                case IParameterSymbol parameter:
-                    if (!_internalInline)
-                        AccessSlot(OpCode.LDARG, _parameters[parameter]);
-                    break;
-                case IPropertySymbol property:
-                    if (property.IsStatic)
-                    {
-                        Call(model, property.GetMethod!);
-                    }
-                    else
-                    {
-                        AddInstruction(OpCode.LDARG0);
-                        Call(model, property.GetMethod!);
-                    }
-                    break;
-                default:
-                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertImplicitArrayCreationExpression(SemanticModel model, ImplicitArrayCreationExpressionSyntax expression)
-        {
-            IArrayTypeSymbol type = (IArrayTypeSymbol)model.GetTypeInfo(expression).ConvertedType!;
-            ConvertInitializerExpression(model, type, expression.Initializer);
-        }
-
-        private void ConvertInitializerExpression(SemanticModel model, InitializerExpressionSyntax expression)
-        {
-            IArrayTypeSymbol type = (IArrayTypeSymbol)model.GetTypeInfo(expression).ConvertedType!;
-            ConvertInitializerExpression(model, type, expression);
-        }
-
-        private void ConvertInitializerExpression(SemanticModel model, IArrayTypeSymbol type, InitializerExpressionSyntax expression)
-        {
-            if (type.ElementType.SpecialType == SpecialType.System_Byte)
-            {
-                Optional<object?>[] values = expression.Expressions.Select(p => model.GetConstantValue(p)).ToArray();
-                if (values.Any(p => !p.HasValue))
-                {
-                    Push(values.Length);
-                    AddInstruction(OpCode.NEWBUFFER);
-                    for (int i = 0; i < expression.Expressions.Count; i++)
-                    {
-                        AddInstruction(OpCode.DUP);
-                        Push(i);
-                        ConvertExpression(model, expression.Expressions[i]);
-                        AddInstruction(OpCode.SETITEM);
-                    }
-                }
-                else
-                {
-                    byte[] data = values.Select(p => (byte)System.Convert.ChangeType(p.Value, typeof(byte))!).ToArray();
-                    Push(data);
-                    ChangeType(VM.Types.StackItemType.Buffer);
-                }
-            }
-            else
-            {
-                for (int i = expression.Expressions.Count - 1; i >= 0; i--)
-                    ConvertExpression(model, expression.Expressions[i]);
-                Push(expression.Expressions.Count);
-                AddInstruction(OpCode.PACK);
-            }
-        }
-
-        private void ConvertInterpolatedStringExpression(SemanticModel model, InterpolatedStringExpressionSyntax expression)
-        {
-            if (expression.Contents.Count == 0)
-            {
-                Push(string.Empty);
-                return;
-            }
-            ConvertInterpolatedStringContent(model, expression.Contents[0]);
-            for (int i = 1; i < expression.Contents.Count; i++)
-            {
-                ConvertInterpolatedStringContent(model, expression.Contents[i]);
-                AddInstruction(OpCode.CAT);
-            }
-            if (expression.Contents.Count >= 2)
-                ChangeType(VM.Types.StackItemType.ByteString);
-        }
-
-        private void ConvertInterpolatedStringContent(SemanticModel model, InterpolatedStringContentSyntax content)
-        {
-            switch (content)
-            {
-                case InterpolatedStringTextSyntax syntax:
-                    Push(syntax.TextToken.ValueText);
-                    break;
-                case InterpolationSyntax syntax:
-                    if (syntax.AlignmentClause is not null)
-                        throw new CompilationException(syntax.AlignmentClause, DiagnosticId.AlignmentClause, $"Alignment clause is not supported: {syntax.AlignmentClause}");
-                    if (syntax.FormatClause is not null)
-                        throw new CompilationException(syntax.FormatClause, DiagnosticId.FormatClause, $"Format clause is not supported: {syntax.FormatClause}");
-                    ConvertObjectToString(model, syntax.Expression);
-                    break;
-            }
-        }
 
         private void ConvertObjectToString(SemanticModel model, ExpressionSyntax expression)
         {
@@ -2254,620 +1011,6 @@ namespace Neo.Compiler
                     break;
                 default:
                     throw new CompilationException(expression, DiagnosticId.InvalidToStringType, $"Unsupported interpolation: {expression}");
-            }
-        }
-
-        private void ConvertInvocationExpression(SemanticModel model, InvocationExpressionSyntax expression)
-        {
-            ArgumentSyntax[] arguments = expression.ArgumentList.Arguments.ToArray();
-            ISymbol symbol = model.GetSymbolInfo(expression.Expression).Symbol!;
-            switch (symbol)
-            {
-                case IEventSymbol @event:
-                    ConvertEventInvocationExpression(model, @event, arguments);
-                    break;
-                case IMethodSymbol method:
-                    ConvertMethodInvocationExpression(model, method, expression.Expression, arguments);
-                    break;
-                default:
-                    ConvertDelegateInvocationExpression(model, expression.Expression, arguments);
-                    break;
-            }
-        }
-
-        private void ConvertEventInvocationExpression(SemanticModel model, IEventSymbol symbol, ArgumentSyntax[] arguments)
-        {
-            AddInstruction(OpCode.NEWARRAY0);
-            foreach (ArgumentSyntax argument in arguments)
-            {
-                AddInstruction(OpCode.DUP);
-                ConvertExpression(model, argument.Expression);
-                AddInstruction(OpCode.APPEND);
-            }
-            Push(symbol.GetDisplayName());
-            Call(ApplicationEngine.System_Runtime_Notify);
-        }
-
-        private void ConvertMethodInvocationExpression(SemanticModel model, IMethodSymbol symbol, ExpressionSyntax expression, ArgumentSyntax[] arguments)
-        {
-            switch (expression)
-            {
-                case IdentifierNameSyntax:
-                    Call(model, symbol, null, arguments);
-                    break;
-                case MemberAccessExpressionSyntax syntax:
-                    if (symbol.IsStatic)
-                        Call(model, symbol, null, arguments);
-                    else
-                        Call(model, symbol, syntax.Expression, arguments);
-                    break;
-                case MemberBindingExpressionSyntax:
-                    Call(model, symbol, true, arguments);
-                    break;
-                default:
-                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported expression: {expression}");
-            }
-        }
-
-        private void ConvertDelegateInvocationExpression(SemanticModel model, ExpressionSyntax expression, ArgumentSyntax[] arguments)
-        {
-            INamedTypeSymbol type = (INamedTypeSymbol)model.GetTypeInfo(expression).Type!;
-            PrepareArgumentsForMethod(model, type.DelegateInvokeMethod!, arguments);
-            ConvertExpression(model, expression);
-            AddInstruction(OpCode.CALLA);
-        }
-
-        private void ConvertIsPatternExpression(SemanticModel model, IsPatternExpressionSyntax expression)
-        {
-            byte anonymousIndex = AddAnonymousVariable();
-            ConvertExpression(model, expression.Expression);
-            AccessSlot(OpCode.STLOC, anonymousIndex);
-            ConvertPattern(model, expression.Pattern, anonymousIndex);
-            RemoveAnonymousVariable(anonymousIndex);
-        }
-
-        private void ConvertMemberAccessExpression(SemanticModel model, MemberAccessExpressionSyntax expression)
-        {
-            ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    if (field.IsConst)
-                    {
-                        Push(field.ConstantValue);
-                    }
-                    else if (field.IsStatic)
-                    {
-                        byte index = context.AddStaticField(field);
-                        AccessSlot(OpCode.LDSFLD, index);
-                    }
-                    else
-                    {
-                        int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                        ConvertExpression(model, expression.Expression);
-                        Push(index);
-                        AddInstruction(OpCode.PICKITEM);
-                    }
-                    break;
-                case IMethodSymbol method:
-                    if (!method.IsStatic)
-                        throw new CompilationException(expression, DiagnosticId.NonStaticDelegate, $"Unsupported delegate: {method}");
-                    MethodConvert convert = context.ConvertMethod(model, method);
-                    Jump(OpCode.PUSHA, convert._startTarget);
-                    break;
-                case IPropertySymbol property:
-                    ExpressionSyntax? instanceExpression = property.IsStatic ? null : expression.Expression;
-                    Call(model, property.GetMethod!, instanceExpression);
-                    break;
-                default:
-                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertMemberBindingExpression(SemanticModel model, MemberBindingExpressionSyntax expression)
-        {
-            ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                    Push(index);
-                    AddInstruction(OpCode.PICKITEM);
-                    break;
-                case IPropertySymbol property:
-                    Call(model, property.GetMethod!);
-                    break;
-                default:
-                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertPostfixUnaryExpression(SemanticModel model, PostfixUnaryExpressionSyntax expression)
-        {
-            switch (expression.OperatorToken.ValueText)
-            {
-                case "++":
-                case "--":
-                    ConvertPostIncrementOrDecrementExpression(model, expression);
-                    break;
-                case "!":
-                    ConvertExpression(model, expression.Operand);
-                    break;
-                default:
-                    throw new CompilationException(expression.OperatorToken, DiagnosticId.SyntaxNotSupported, $"Unsupported operator: {expression.OperatorToken}");
-            }
-        }
-
-        private void ConvertPostIncrementOrDecrementExpression(SemanticModel model, PostfixUnaryExpressionSyntax expression)
-        {
-            switch (expression.Operand)
-            {
-                case ElementAccessExpressionSyntax operand:
-                    ConvertElementAccessPostIncrementOrDecrementExpression(model, expression.OperatorToken, operand);
-                    break;
-                case IdentifierNameSyntax operand:
-                    ConvertIdentifierNamePostIncrementOrDecrementExpression(model, expression.OperatorToken, operand);
-                    break;
-                case MemberAccessExpressionSyntax operand:
-                    ConvertMemberAccessPostIncrementOrDecrementExpression(model, expression.OperatorToken, operand);
-                    break;
-                default:
-                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported postfix unary expression: {expression}");
-            }
-        }
-
-        private void ConvertElementAccessPostIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, ElementAccessExpressionSyntax operand)
-        {
-            if (operand.ArgumentList.Arguments.Count != 1)
-                throw new CompilationException(operand.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {operand.ArgumentList.Arguments}");
-            if (model.GetSymbolInfo(operand).Symbol is IPropertySymbol property)
-            {
-                ConvertExpression(model, operand.Expression);
-                ConvertExpression(model, operand.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                Call(model, property.GetMethod!, CallingConvention.StdCall);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                AddInstruction(OpCode.REVERSE3);
-                EmitIncrementOrDecrement(operatorToken, property.Type);
-                Call(model, property.SetMethod!, CallingConvention.StdCall);
-            }
-            else
-            {
-                ConvertExpression(model, operand.Expression);
-                ConvertExpression(model, operand.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.PICKITEM);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                AddInstruction(OpCode.REVERSE3);
-                EmitIncrementOrDecrement(operatorToken, model.GetTypeInfo(operand).Type);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertIdentifierNamePostIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, IdentifierNameSyntax operand)
-        {
-            ISymbol symbol = model.GetSymbolInfo(operand).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldIdentifierNamePostIncrementOrDecrementExpression(operatorToken, field);
-                    break;
-                case ILocalSymbol local:
-                    ConvertLocalIdentifierNamePostIncrementOrDecrementExpression(operatorToken, local);
-                    break;
-                case IParameterSymbol parameter:
-                    ConvertParameterIdentifierNamePostIncrementOrDecrementExpression(operatorToken, parameter);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyIdentifierNamePostIncrementOrDecrementExpression(model, operatorToken, property);
-                    break;
-                default:
-                    throw new CompilationException(operand, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldIdentifierNamePostIncrementOrDecrementExpression(SyntaxToken operatorToken, IFieldSymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                byte index = context.AddStaticField(symbol);
-                AccessSlot(OpCode.LDSFLD, index);
-                AddInstruction(OpCode.DUP);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(symbol.ContainingType.GetFields(), symbol);
-                AddInstruction(OpCode.LDARG0);
-                AddInstruction(OpCode.DUP);
-                Push(index);
-                AddInstruction(OpCode.PICKITEM);
-                AddInstruction(OpCode.TUCK);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                Push(index);
-                AddInstruction(OpCode.SWAP);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertLocalIdentifierNamePostIncrementOrDecrementExpression(SyntaxToken operatorToken, ILocalSymbol symbol)
-        {
-            byte index = _localVariables[symbol];
-            AccessSlot(OpCode.LDLOC, index);
-            AddInstruction(OpCode.DUP);
-            EmitIncrementOrDecrement(operatorToken, symbol.Type);
-            AccessSlot(OpCode.STLOC, index);
-        }
-
-        private void ConvertParameterIdentifierNamePostIncrementOrDecrementExpression(SyntaxToken operatorToken, IParameterSymbol symbol)
-        {
-            byte index = _parameters[symbol];
-            AccessSlot(OpCode.LDARG, index);
-            AddInstruction(OpCode.DUP);
-            EmitIncrementOrDecrement(operatorToken, symbol.Type);
-            AccessSlot(OpCode.STARG, index);
-        }
-
-        private void ConvertPropertyIdentifierNamePostIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, IPropertySymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                Call(model, symbol.GetMethod!);
-                AddInstruction(OpCode.DUP);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                Call(model, symbol.SetMethod!);
-            }
-            else
-            {
-                AddInstruction(OpCode.LDARG0);
-                AddInstruction(OpCode.DUP);
-                Call(model, symbol.GetMethod!);
-                AddInstruction(OpCode.TUCK);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                Call(model, symbol.SetMethod!, CallingConvention.StdCall);
-            }
-        }
-
-        private void ConvertMemberAccessPostIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, MemberAccessExpressionSyntax operand)
-        {
-            ISymbol symbol = model.GetSymbolInfo(operand).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldMemberAccessPostIncrementOrDecrementExpression(model, operatorToken, operand, field);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyMemberAccessPostIncrementOrDecrementExpression(model, operatorToken, operand, property);
-                    break;
-                default:
-                    throw new CompilationException(operand, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldMemberAccessPostIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, MemberAccessExpressionSyntax operand, IFieldSymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                byte index = context.AddStaticField(symbol);
-                AccessSlot(OpCode.LDSFLD, index);
-                AddInstruction(OpCode.DUP);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(symbol.ContainingType.GetFields(), symbol);
-                ConvertExpression(model, operand.Expression);
-                AddInstruction(OpCode.DUP);
-                Push(index);
-                AddInstruction(OpCode.PICKITEM);
-                AddInstruction(OpCode.TUCK);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                Push(index);
-                AddInstruction(OpCode.SWAP);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertPropertyMemberAccessPostIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, MemberAccessExpressionSyntax operand, IPropertySymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                Call(model, symbol.GetMethod!);
-                AddInstruction(OpCode.DUP);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                Call(model, symbol.SetMethod!);
-            }
-            else
-            {
-                ConvertExpression(model, operand.Expression);
-                AddInstruction(OpCode.DUP);
-                Call(model, symbol.GetMethod!);
-                AddInstruction(OpCode.TUCK);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                Call(model, symbol.SetMethod!, CallingConvention.StdCall);
-            }
-        }
-
-        private void ConvertPrefixUnaryExpression(SemanticModel model, PrefixUnaryExpressionSyntax expression)
-        {
-            switch (expression.OperatorToken.ValueText)
-            {
-                case "+":
-                    ConvertExpression(model, expression.Operand);
-                    break;
-                case "-":
-                    ConvertExpression(model, expression.Operand);
-                    AddInstruction(OpCode.NEGATE);
-                    break;
-                case "~":
-                    ConvertExpression(model, expression.Operand);
-                    AddInstruction(OpCode.INVERT);
-                    break;
-                case "!":
-                    ConvertExpression(model, expression.Operand);
-                    AddInstruction(OpCode.NOT);
-                    break;
-                case "++":
-                case "--":
-                    ConvertPreIncrementOrDecrementExpression(model, expression);
-                    break;
-                case "^":
-                    AddInstruction(OpCode.DUP);
-                    AddInstruction(OpCode.SIZE);
-                    ConvertExpression(model, expression.Operand);
-                    AddInstruction(OpCode.SUB);
-                    break;
-                default:
-                    throw new CompilationException(expression.OperatorToken, DiagnosticId.SyntaxNotSupported, $"Unsupported operator: {expression.OperatorToken}");
-            }
-        }
-
-        private void ConvertPreIncrementOrDecrementExpression(SemanticModel model, PrefixUnaryExpressionSyntax expression)
-        {
-            switch (expression.Operand)
-            {
-                case ElementAccessExpressionSyntax operand:
-                    ConvertElementAccessPreIncrementOrDecrementExpression(model, expression.OperatorToken, operand);
-                    break;
-                case IdentifierNameSyntax operand:
-                    ConvertIdentifierNamePreIncrementOrDecrementExpression(model, expression.OperatorToken, operand);
-                    break;
-                case MemberAccessExpressionSyntax operand:
-                    ConvertMemberAccessPreIncrementOrDecrementExpression(model, expression.OperatorToken, operand);
-                    break;
-                default:
-                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported postfix unary expression: {expression}");
-            }
-        }
-
-        private void ConvertElementAccessPreIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, ElementAccessExpressionSyntax operand)
-        {
-            if (operand.ArgumentList.Arguments.Count != 1)
-                throw new CompilationException(operand.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {operand.ArgumentList.Arguments}");
-            if (model.GetSymbolInfo(operand).Symbol is IPropertySymbol property)
-            {
-                ConvertExpression(model, operand.Expression);
-                ConvertExpression(model, operand.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                Call(model, property.GetMethod!, CallingConvention.StdCall);
-                EmitIncrementOrDecrement(operatorToken, property.Type);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                Call(model, property.SetMethod!, CallingConvention.Cdecl);
-            }
-            else
-            {
-                ConvertExpression(model, operand.Expression);
-                ConvertExpression(model, operand.ArgumentList.Arguments[0].Expression);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.OVER);
-                AddInstruction(OpCode.PICKITEM);
-                EmitIncrementOrDecrement(operatorToken, model.GetTypeInfo(operand).Type);
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.REVERSE4);
-                AddInstruction(OpCode.REVERSE3);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertIdentifierNamePreIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, IdentifierNameSyntax operand)
-        {
-            ISymbol symbol = model.GetSymbolInfo(operand).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldIdentifierNamePreIncrementOrDecrementExpression(operatorToken, field);
-                    break;
-                case ILocalSymbol local:
-                    ConvertLocalIdentifierNamePreIncrementOrDecrementExpression(operatorToken, local);
-                    break;
-                case IParameterSymbol parameter:
-                    ConvertParameterIdentifierNamePreIncrementOrDecrementExpression(operatorToken, parameter);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyIdentifierNamePreIncrementOrDecrementExpression(model, operatorToken, property);
-                    break;
-                default:
-                    throw new CompilationException(operand, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldIdentifierNamePreIncrementOrDecrementExpression(SyntaxToken operatorToken, IFieldSymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                byte index = context.AddStaticField(symbol);
-                AccessSlot(OpCode.LDSFLD, index);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(symbol.ContainingType.GetFields(), symbol);
-                AddInstruction(OpCode.LDARG0);
-                AddInstruction(OpCode.DUP);
-                Push(index);
-                AddInstruction(OpCode.PICKITEM);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.TUCK);
-                Push(index);
-                AddInstruction(OpCode.SWAP);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertLocalIdentifierNamePreIncrementOrDecrementExpression(SyntaxToken operatorToken, ILocalSymbol symbol)
-        {
-            byte index = _localVariables[symbol];
-            AccessSlot(OpCode.LDLOC, index);
-            EmitIncrementOrDecrement(operatorToken, symbol.Type);
-            AddInstruction(OpCode.DUP);
-            AccessSlot(OpCode.STLOC, index);
-        }
-
-        private void ConvertParameterIdentifierNamePreIncrementOrDecrementExpression(SyntaxToken operatorToken, IParameterSymbol symbol)
-        {
-            byte index = _parameters[symbol];
-            AccessSlot(OpCode.LDARG, index);
-            EmitIncrementOrDecrement(operatorToken, symbol.Type);
-            AddInstruction(OpCode.DUP);
-            AccessSlot(OpCode.STARG, index);
-        }
-
-        private void ConvertPropertyIdentifierNamePreIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, IPropertySymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                Call(model, symbol.GetMethod!);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.DUP);
-                Call(model, symbol.SetMethod!);
-            }
-            else
-            {
-                AddInstruction(OpCode.LDARG0);
-                AddInstruction(OpCode.DUP);
-                Call(model, symbol.GetMethod!);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.TUCK);
-                Call(model, symbol.SetMethod!, CallingConvention.StdCall);
-            }
-        }
-
-        private void ConvertMemberAccessPreIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, MemberAccessExpressionSyntax operand)
-        {
-            ISymbol symbol = model.GetSymbolInfo(operand).Symbol!;
-            switch (symbol)
-            {
-                case IFieldSymbol field:
-                    ConvertFieldMemberAccessPreIncrementOrDecrementExpression(model, operatorToken, operand, field);
-                    break;
-                case IPropertySymbol property:
-                    ConvertPropertyMemberAccessPreIncrementOrDecrementExpression(model, operatorToken, operand, property);
-                    break;
-                default:
-                    throw new CompilationException(operand, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-            }
-        }
-
-        private void ConvertFieldMemberAccessPreIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, MemberAccessExpressionSyntax operand, IFieldSymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                byte index = context.AddStaticField(symbol);
-                AccessSlot(OpCode.LDSFLD, index);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STSFLD, index);
-            }
-            else
-            {
-                int index = Array.IndexOf(symbol.ContainingType.GetFields(), symbol);
-                ConvertExpression(model, operand.Expression);
-                AddInstruction(OpCode.DUP);
-                Push(index);
-                AddInstruction(OpCode.PICKITEM);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.TUCK);
-                Push(index);
-                AddInstruction(OpCode.SWAP);
-                AddInstruction(OpCode.SETITEM);
-            }
-        }
-
-        private void ConvertPropertyMemberAccessPreIncrementOrDecrementExpression(SemanticModel model, SyntaxToken operatorToken, MemberAccessExpressionSyntax operand, IPropertySymbol symbol)
-        {
-            if (symbol.IsStatic)
-            {
-                Call(model, symbol.GetMethod!);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.DUP);
-                Call(model, symbol.SetMethod!);
-            }
-            else
-            {
-                ConvertExpression(model, operand.Expression);
-                AddInstruction(OpCode.DUP);
-                Call(model, symbol.GetMethod!);
-                EmitIncrementOrDecrement(operatorToken, symbol.Type);
-                AddInstruction(OpCode.TUCK);
-                Call(model, symbol.SetMethod!, CallingConvention.StdCall);
-            }
-        }
-
-        private void EmitIncrementOrDecrement(SyntaxToken operatorToken, ITypeSymbol? typeSymbol)
-        {
-            AddInstruction(operatorToken.ValueText switch
-            {
-                "++" => OpCode.INC,
-                "--" => OpCode.DEC,
-                _ => throw new CompilationException(operatorToken, DiagnosticId.SyntaxNotSupported, $"Unsupported operator: {operatorToken}")
-            });
-            if (typeSymbol != null) EnsureIntegerInRange(typeSymbol);
-        }
-
-        private void ConvertSwitchExpression(SemanticModel model, SwitchExpressionSyntax expression)
-        {
-            var arms = expression.Arms.Select(p => (p, new JumpTarget())).ToArray();
-            JumpTarget breakTarget = new();
-            byte anonymousIndex = AddAnonymousVariable();
-            ConvertExpression(model, expression.GoverningExpression);
-            AccessSlot(OpCode.STLOC, anonymousIndex);
-            foreach (var (arm, nextTarget) in arms)
-            {
-                ConvertPattern(model, arm.Pattern, anonymousIndex);
-                Jump(OpCode.JMPIFNOT_L, nextTarget);
-                if (arm.WhenClause is not null)
-                {
-                    ConvertExpression(model, arm.WhenClause.Condition);
-                    Jump(OpCode.JMPIFNOT_L, nextTarget);
-                }
-                ConvertExpression(model, arm.Expression);
-                Jump(OpCode.JMP_L, breakTarget);
-                nextTarget.Instruction = AddInstruction(OpCode.NOP);
-            }
-            AccessSlot(OpCode.LDLOC, anonymousIndex);
-            AddInstruction(OpCode.THROW);
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            RemoveAnonymousVariable(anonymousIndex);
-        }
-
-        private void ConvertTupleExpression(SemanticModel model, TupleExpressionSyntax expression)
-        {
-            AddInstruction(OpCode.NEWSTRUCT0);
-            foreach (ArgumentSyntax argument in expression.Arguments)
-            {
-                AddInstruction(OpCode.DUP);
-                ConvertExpression(model, argument.Expression);
-                AddInstruction(OpCode.APPEND);
             }
         }
